Deliverable (9) - Python Data Analytics
Following the model's evaluation, we discovered that the decision tree classifier provided us with an accuracy of 0.80, meaning that the model correctly 
predicted the survival status of 80% of the passengers. The precision and recall of our model were 0.77 and 0.71, respectively, indicating moderate accuracy.
As a result, we were successful in creating a decision tree classifier to forecast passenger survivability based on the Titanic dataset's provided attributes. 
The model can be used to forecast whether new passengers will survive.